# shared/schemas/api.openapi.json
{
  "components": {
    "schemas": {
      "AbendContext": {
        "properties": {
          "job_name": {
            "title": "Job Name",
            "type": "string"
          },
          "occurred_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At"
          },
          "region": {
            "title": "Region",
            "type": "string"
          }
        },
        "required": [
          "region",
          "job_name"
        ],
        "title": "AbendContext",
        "type": "object"
      },
...
      "AbendRequest": {
        "properties": {
          "context": {
            "$ref": "#/components/schemas/AbendContext"
          },
          "raw_text": {
            "minLength": 1,
            "title": "Raw Text",
            "type": "string"
          }
        },
        "required": [
          "raw_text",
          "context"
        ],
        "title": "AbendRequest",
        "type": "object"
      },
      "AbendResolveRequest": {
        "properties": {
          "applied_runbook_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied Runbook Id"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "outcome": {
            "title": "Outcome",
            "type": "string"
          },
          "root_cause_choice": {
            "title": "Root Cause Choice",
            "type": "string"
          }
        },
        "required": [
          "root_cause_choice",
          "outcome"
        ],
        "title": "AbendResolveRequest",
        "type": "object"
      },
      "AbendResolveResponse": {
        "properties": {
          "audit_event_id": {
            "title": "Audit Event Id",
            "type": "string"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "learning_event_id": {
            "title": "Learning Event Id",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "audit_event_id",
          "learning_event_id"
        ],
        "title": "AbendResolveResponse",
        "type": "object"
      },
      "AbendResponse": {
        "additionalProperties": true,
        "description": "Full ABEND analysis envelope. ``degraded`` carries the unfamiliar-tier flow.",
        "properties": {
          "business_rule_explanation": {
            "title": "Business Rule Explanation",
            "type": "string"
          },
          "degraded": {
            "default": false,
            "title": "Degraded",
            "type": "boolean"
          },
          "degraded_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Degraded Reason"
          },
          "failing_step": {
            "$ref": "#/components/schemas/FailingStep"
          },
          "identified_abend": {
            "$ref": "#/components/schemas/IdentifiedAbend"
          },
          "matching_runbooks": {
            "items": {
              "$ref": "#/components/schemas/MatchingRunbook"
            },
            "title": "Matching Runbooks",
            "type": "array"
          },
          "ranked_root_causes": {
            "items": {
              "$ref": "#/components/schemas/RankedRootCause"
            },
            "title": "Ranked Root Causes",
            "type": "array"
          },
          "source_trace": {
            "$ref": "#/components/schemas/SourceTrace"
          }
        },
        "required": [
          "identified_abend",
          "failing_step",
          "source_trace",
          "business_rule_explanation",
          "ranked_root_causes",
          "matching_runbooks"
        ],
        "title": "AbendResponse",
        "type": "object"
      },
...
      "AttestationResponse": {
        "properties": {
          "chain_root_hash": {
            "title": "Chain Root Hash",
            "type": "string"
          },
          "date": {
            "title": "Date",
            "type": "string"
          },
          "event_count": {
            "title": "Event Count",
            "type": "integer"
          }
        },
        "required": [
          "date",
          "chain_root_hash",
          "event_count"
        ],
        "title": "AttestationResponse",
        "type": "object"
      },
      "AuditBundleRequest": {
        "properties": {
          "actor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor"
          },
          "from": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From"
          },
          "subject_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Kind"
          },
          "subject_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Name"
          },
          "subject_region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Region"
          },
          "to": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "To"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          }
        },
        "title": "AuditBundleRequest",
        "type": "object"
      },
      "AuditBundleResponse": {
        "properties": {
          "bundle_job_id": {
            "title": "Bundle Job Id",
            "type": "string"
          }
        },
        "required": [
          "bundle_job_id"
        ],
        "title": "AuditBundleResponse",
        "type": "object"
      },
      "AuditBundleStatusResponse": {
        "properties": {
          "bundle_job_id": {
            "title": "Bundle Job Id",
            "type": "string"
          },
          "download_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Url"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "bundle_job_id",
          "state"
        ],
        "title": "AuditBundleStatusResponse",
        "type": "object"
      },
...
      "AuditQueryResponse": {
        "properties": {
          "bookmark": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bookmark"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/AuditEvent"
            },
            "title": "Events",
            "type": "array"
          }
        },
        "required": [
          "events"
        ],
        "title": "AuditQueryResponse",
        "type": "object"
      },
...
      "Envelope_AbendResolveResponse_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AbendResolveResponse"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[AbendResolveResponse]",
        "type": "object"
      },
      "Envelope_AbendResponse_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AbendResponse"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[AbendResponse]",
        "type": "object"
      },
      "Envelope_AttestationResponse_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AttestationResponse"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[AttestationResponse]",
        "type": "object"
      },
      "Envelope_AuditBundleResponse_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AuditBundleResponse"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[AuditBundleResponse]",
        "type": "object"
      },
      "Envelope_AuditBundleStatusResponse_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AuditBundleStatusResponse"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[AuditBundleStatusResponse]",
        "type": "object"
      },
      "Envelope_AuditEvent_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AuditEvent"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[AuditEvent]",
        "type": "object"
      },
      "Envelope_AuditQueryResponse_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AuditQueryResponse"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[AuditQueryResponse]",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MatchingRunbook": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "success_count": {
            "title": "Success Count",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "success_count"
        ],
        "title": "MatchingRunbook",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "AI control plane for z/OS modernization. See docs/API.md.",
    "title": "Helios Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/abend": {
      "post": {
        "operationId": "analyze_abend_api_abend_post",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AbendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AbendResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze Abend",
        "tags": [
          "abend"
        ]
      }
    },
    "/api/abend/{event_id}/resolve": {
      "post": {
        "operationId": "resolve_abend_api_abend__event_id__resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AbendResolveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AbendResolveResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Abend",
        "tags": [
          "abend"
        ]
      }
    },
    "/api/audit": {
      "get": {
        "operationId": "query_audit_api_audit_get",
        "parameters": [
          {
            "in": "query",
            "name": "subject_kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subject Kind"
            }
          },
          {
            "in": "query",
            "name": "subject_name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subject Name"
            }
          },
          {
            "in": "query",
            "name": "actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Actor"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AuditQueryResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Query Audit",
        "tags": [
          "audit"
        ]
      }
    },
    "/api/audit/attestation/{date}": {
      "get": {
        "description": "Read the daily chain root attestation written by ``audit_chain_verify``.",
        "operationId": "attestation_api_audit_attestation__date__get",
        "parameters": [
          {
            "in": "path",
            "name": "date",
            "required": true,
            "schema": {
              "title": "Date",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AttestationResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Attestation",
        "tags": [
          "audit"
        ]
      }
    },
    "/api/audit/bundle": {
      "post": {
        "operationId": "request_bundle_api_audit_bundle_post",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuditBundleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AuditBundleResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request Bundle",
        "tags": [
          "audit"
        ]
      }
    },
    "/api/audit/bundle/{job_id}": {
      "get": {
        "operationId": "bundle_status_api_audit_bundle__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AuditBundleStatusResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bundle Status",
        "tags": [
          "audit"
        ]
      }
    },
    "/api/audit/{event_id}": {
      "get": {
        "operationId": "get_audit_event_api_audit__event_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AuditEvent_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Audit Event",
        "tags": [
          "audit"
        ]
      }
    },
...
  }
}